+++ server-side/src/main/java/com/tick42/quicksilver/repositories/base/GitHubRepository.java
@@ -3,6 +3,6 @@ package com.tick42.quicksilver.repositories.base;
 import com.tick42.quicksilver.models.GitHubModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GitHubRepository extends JpaRepository<GitHubModel, Integer> {
+public interface GitHubRepository extends JpaRepository<GitHubModel, Long> {
 
 }
