+++ server-side/src/main/java/com/tick42/quicksilver/repositories/base/SettingsRepository.java
@@ -4,6 +4,6 @@ import com.tick42.quicksilver.models.Settings;
 import com.tick42.quicksilver.models.UserModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SettingsRepository extends JpaRepository<Settings, Integer> {
+public interface SettingsRepository extends JpaRepository<Settings, Long> {
     Settings findByUser(UserModel user);
 }
